+++ apps/web-app/src/features/nodes/client/tree/mutations/index.ts
@@ -1,6 +1,5 @@
 export { useCreateMutation } from "./use-create-node";
 export { useDuplicateMutation } from "./use-duplicate-node";
-export { useLoadMoreMutation } from "./use-load-more-nodes";
 export { useMoveMutation } from "./use-move-node";
 export { useRemoveMutation } from "./use-remove-node";
 export { useSetDueDateMutation } from "./use-set-node-due-date";